Unit test template updates.
