--- tooling/BRIck Integration Tests.xctemplate/TemplateInfo.plist
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>Kind</key>
-	<string>Xcode.IDEFoundation.TextSubstitutionFileTemplateKind</string>
-	<key>Summary</key>
-	<string>BRIck Integration Tests using Swift Testing</string>
-	<key>Description</key>
-	<string>BRIck integration test files.</string>
-	<key>SortOrder</key>
-	<string>2</string>
-	<key>AllowedTypes</key>
-	<dict>
-		<key>Item 0</key>
-		<string>public.swift-source</string>
-	</dict>
-	<key>DefaultCompletionName</key>
-	<string>BRIck Integration Tests</string>
-	<key>MainTemplateFiles</key>
-	<string>___FILEBASENAME___.swift</string>
-	<key>Platforms</key>
-	<array/>
-	<key>Options</key>
-	<array>
-		<dict>
-			<key>Identifier</key>
-			<string>productName</string>
-			<key>Required</key>
-			<string>true</string>
-			<key>Name</key>
-			<string>Name of the BRIck to test:</string>
-			<key>Description</key>
-			<string>The name of the BRIck to create integration test for.</string>
-			<key>Type</key>
-			<string>text</string>
-			<key>NotPersisted</key>
-			<string>true</string>
-		</dict>
-	</array>
-</dict>
-</plist>
