--- tooling/BRIck Unit Tests.xctemplate/___FILEBASENAME___InteractorUnitTests.swift
@@ -1,16 +1,0 @@
-// ___FILEHEADER___
-
-@testable import ___PROJECTNAME___
-import Testing
-
-final class ___VARIABLE_productName___InteractorUnitTests {
-    private var interactor: ___VARIABLE_productName___Interactor!
-
-    // TODO: declare other objects and mocks you need as provate vars
-
-    init() {}
-
-    deinit { interactor = nil }
-
-    // MARK: - Tests
-}
