--- tooling/BRIck Unit Tests.xctemplate/___FILEBASENAME___RouterUnitTests.swift
@@ -1,20 +1,0 @@
-// ___FILEHEADER___
-
-@testable import ___PROJECTNAME___
-import SwiftUI
-import Testing
-
-final class ___VARIABLE_productName___RouterUnitTests {
-    private var router: ___VARIABLE_productName___Router<
-        ___VARIABLE_productName___Presenter,
-        EmptyView
-    >!
-
-    // TODO: declare others objects and mocks you need as private vars
-
-    init() {}
-
-    deinit { router = nil }
-
-    // MARK: - Tests
-}
