--- tooling/BRIck Unit XCTests.xctemplate/___FILEBASENAME___InteractorUnitTests.swift
@@ -1,22 +1,0 @@
-// ___FILEHEADER___
-
-@testable import ___PROJECTNAME___
-import XCTest
-
-final class ___VARIABLE_productName___InteractorUnitTests: XCTestCase {
-    private var interactor: ___VARIABLE_productName___Interactor!
-
-    // TODO: declare other objects and mocks you need as provate vars
-
-    override func setUp() {
-        super.setUp()
-    }
-
-    override func tearDown() {
-        super.tearDown()
-
-        interactor = nil
-    }
-
-    // MARK: - Tests
-}
