--- tooling/BRIck Unit XCTests.xctemplate/___FILEBASENAME___RouterUnitTests.swift
@@ -1,26 +1,0 @@
-// ___FILEHEADER___
-
-@testable import ___PROJECTNAME___
-import SwiftUI
-import XCTest
-
-final class ___VARIABLE_productName___RouterUnitTests: XCTestCase {
-    private var router: ___VARIABLE_productName___Router<
-        ___VARIABLE_productName___Presenter,
-        EmptyView
-    >!
-
-    // TODO: declare others objects and mocks you need as private vars
-
-    override func setUp() {
-        super.setUp()
-    }
-
-    override func tearDown() {
-        super.tearDown()
-
-        router = nil
-    }
-
-    // MARK: - Tests
-}
